plot.mvsusy {mvSUSY}	R Documentation

mvsusy plot method

Description

Generate plot for a mvsusy object.

Usage

  ## S3 method for class 'mvsusy'
plot(x,
    type=c("eigenvalue","density","free scale","segment-wise","time series"),
    ...,
    bins,
    plotly)

Arguments

x
A mvsusy object.

type
Character to specify type of the plot, one of eigenvalue, density, free scale, segment-wise, time series.

...
Ignored.

bins
Numeric. Number of bins used in free scale plot type, by default number of segments divided by 2.

plotly
Logical. Passing FALSE will force time series plot not to use plotly even if installed.

Details

Eigenvalue plot type works only for mvSUSY computed using lambda_max method. Time series plot will by default use plotly package, if installed, to provide interactive plot.

Value

Returns NULL invisibly. Generate plot as a side effect.

See Also

mvsusy

Examples

set.seed(1)
data = as.data.frame(replicate(5, sample(10, 5000, TRUE)))
res = mvsusy(data, segment=10, Hz=10)
plot(res, type="eigenvalue")
plot(res, type="density")
plot(res, type="free scale")
plot(res, type="segment-wise")
plot(res, type="time series")
plot(res, type="time series", plotly=FALSE)

[Package mvSUSY version 0.1.1 Index]